Refactor authentication logic by modularizing `initAuth` functionality into separate files.
