+++ src/lib/auth/init/fetchCurrentUser.ts
@@ -1,9 +1,9 @@
-import { CurrentUserDocument, GqlCurrentUserPayload } from "@/types/graphql";
+import { CurrentUserServerDocument, GqlCurrentUserPayload } from "@/types/graphql";
 import { apolloClient } from "@/lib/apollo";
 
 export async function fetchCurrentUserClient(): Promise<GqlCurrentUserPayload["user"] | null> {
   const { data } = await apolloClient.query({
-    query: CurrentUserDocument,
+    query: CurrentUserServerDocument,
     fetchPolicy: "network-only",
   });
   return data?.currentUser?.user ?? null;
